pyproject.toml
[tool.poetry]
name = "vyapar"
version = "0.1.0"
description = ""
authors = ["Aditya Kumar Kasaudhan <akasaudhan02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pylint = "3.3.0"
asgiref = "3.8.1"
astroid = "3.3.3"
dependency-injector = "4.42.0"
dill = "0.3.8"
django = "5.1.1"
django-environ = "0.11.2"
django-log-request-id = "2.1.0"
djangorestframework = "3.15.2"
dnspython = "2.6.1"
email-validator = "1.3.1"
idna = "3.10"
isort = "5.13.2"
mccabe = "0.7.0"
orjson = "3.10.7"
platformdirs = "4.3.6"
psycopg2-binary = "2.9.9"
pydantic = "1.10.2"
six = "1.16.0"
sqlparse = "0.5.1"
tomlkit = "0.13.2"
typing-extensions = "4.12.2"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.djlint]
profile="django"

[tool.ruff]
line-length = 120
